+++ src/components/Main/MainView/ChannelView/ChannelViewContent/ChannelViewContentMain.vue
@@ -11,6 +11,7 @@
       :last-loading-direction="lastLoadingDirection"
       @request-load-former="onLoadFormerMessagesRequest"
       @request-load-latter="onLoadLatterMessagesRequest"
+      @reset-is-reached-latest="resetIsReachedLatest"
     >
       <template #default="{ messageId, onChangeHeight, onEntryMessageLoaded }">
         <messages-scroller-separator
@@ -100,6 +101,10 @@ const isArchived = computed(
 const messagePinnedUserMap = computed(
   () => new Map(props.pinnedMessages.map(pin => [pin.message.id, pin.userId]))
 )
+
+const resetIsReachedLatest = () => {
+  isReachedLatest.value = false
+}
 </script>
 
 <style lang="scss" module>
